fix: split oversized preview tab files under the max-lines ratchet
